🎨style: implement hide header feature
