+++ app/globals.css
@@ -2,51 +2,66 @@
 @import "tailwindcss/components";
 @import "tailwindcss/utilities";
 
+.fade-in-up {
+  animation: fade-in-up 1.5s;
+}
+
 .tilt-shaking:hover {
-animation: tilt-shaking 0.25s linear infinite;
+  animation: tilt-shaking 0.25s linear infinite;
 }
 
 .typing {
-animation: typing 2s steps(20), blink .7s infinite;
-border-right: .15em solid orange;
-margin: auto;
-overflow: hidden;
-white-space: nowrap;
+  animation: typing 2s steps(20), blink .7s infinite;
+  border-right: .15em solid orange;
+  margin: auto;
+  overflow: hidden;
+  white-space: nowrap;
+}
+
+@keyframes fade-in-up {
+  from {
+    opacity: 0;
+    transform: translatey(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translatey(0);
+  }
 }
 
 @keyframes tilt-shaking {
-0% {
-transform: rotate(0deg);
-}
-25% {
-transform: rotate(5deg);
-}
-50% {
-transform: rotate(0eg);
-}
-75% {
-transform: rotate(-5deg);
-}
-100% {
-transform: rotate(0deg);
-}
+  0% {
+    transform: rotate(0deg);
+  }
+  25% {
+    transform: rotate(5deg);
+  }
+  50% {
+    transform: rotate(0eg);
+  }
+  75% {
+    transform: rotate(-5deg);
+  }
+  100% {
+    transform: rotate(0deg);
+  }
 }
 
 @keyframes blink {
-from,
-to {
-border-color: transparent;
-}
-50% {
-border-color: orange;
-}
+  from,
+  to {
+    border-color: transparent;
+  }
+  50% {
+    border-color: orange;
+  }
 }
 
 @keyframes typing {
-0% {
-width: 0%;
-}
-100% {
-width: 100%;
-}
+  0% {
+    width: 0%;
+  }
+  100% {
+    width: 100%;
+  }
 }
