+++ biome.json
@@ -32,7 +32,8 @@
     },
     "formatter": {
       "enabled": true,
-      "indentStyle": "space"
+      "indentStyle": "space",
+      "indentWidth": 2
     },
     "linter": {
       "enabled": true
